Factorial

+++ factorial.c
@@ -4,12 +4,12 @@ float factorialF(float num)
 {
 	if(num == 0)
 		return 1;
-	else if(num < TABLE_SIZE && table[num] > 0)
-		return table[num];
+	else if(num < TABLE_SIZE && table[(int)num] > 0)
+		return table[(int)num];
 	else
 	{
 		float value = factorialF(num - 1) * num;
-		table[num] = value;
+		table[(int)num] = value;
 		return value;
 	}
 }
